[Alpha WIP] Open Factory SqliteConnections (#30)
Return SqliteConnection from open factory for web
