+++ lib/src/impl/stub_sqlite_open_factory.dart
@@ -1,5 +1,8 @@
+import 'dart:async';
+
 import 'package:sqlite_async/sqlite3_common.dart';
 import 'package:sqlite_async/src/common/abstract_open_factory.dart';
+import 'package:sqlite_async/src/sqlite_connection.dart';
 import 'package:sqlite_async/src/sqlite_options.dart';
 
 class DefaultSqliteOpenFactory extends AbstractDefaultSqliteOpenFactory {
@@ -14,6 +17,11 @@ class DefaultSqliteOpenFactory extends AbstractDefaultSqliteOpenFactory {
 
   @override
   List<String> pragmaStatements(SqliteOpenOptions options) {
+    throw UnimplementedError();
+  }
+
+  @override
+  FutureOr<SqliteConnection> openConnection(SqliteOpenOptions options) {
     throw UnimplementedError();
   }
 }
