+++ lib/src/native/database/native_sqlite_database.dart
@@ -1,9 +1,8 @@
 import 'dart:async';
-import 'dart:isolate';
 
+import 'package:meta/meta.dart';
 import 'package:sqlite_async/src/common/abstract_open_factory.dart';
 import 'package:sqlite_async/src/common/sqlite_database.dart';
-import 'package:sqlite_async/src/common/port_channel.dart';
 import 'package:sqlite_async/src/native/database/connection_pool.dart';
 import 'package:sqlite_async/src/native/database/native_sqlite_connection_impl.dart';
 import 'package:sqlite_async/src/native/native_isolate_connection_factory.dart';
@@ -13,8 +12,6 @@ import 'package:sqlite_async/src/sqlite_connection.dart';
 import 'package:sqlite_async/src/sqlite_options.dart';
 import 'package:sqlite_async/src/sqlite_queries.dart';
 import 'package:sqlite_async/src/update_notification.dart';
-import 'package:sqlite_async/src/utils/native_database_utils.dart';
-import 'package:sqlite_async/src/utils/shared_utils.dart';
 
 /// A SQLite database instance.
 ///
@@ -24,7 +21,7 @@ class SqliteDatabaseImpl
     with SqliteQueries, SqliteDatabaseMixin
     implements SqliteDatabase {
   @override
-  final AbstractDefaultSqliteOpenFactory openFactory;
+  final DefaultSqliteOpenFactory openFactory;
 
   @override
   late Stream<UpdateNotification> updates;
@@ -32,16 +29,16 @@ class SqliteDatabaseImpl
   @override
   int maxReaders;
 
-  @override
-  late Future<void> isInitialized;
+  /// Global lock to serialize write transactions.
+  final SimpleMutex mutex = SimpleMutex();
 
-  late final PortServer _eventsPort;
+  @override
+  @protected
+  // Native doesn't require any asynchronous initialization
+  late Future<void> isInitialized = Future.value();
 
   late final SqliteConnectionImpl _internalConnection;
   late final SqliteConnectionPool _pool;
-
-  /// Global lock to serialize write transactions.
-  final SimpleMutex mutex = SimpleMutex();
 
   /// Open a SqliteDatabase.
   ///
@@ -71,26 +68,17 @@ class SqliteDatabaseImpl
   ///  2. Running additional per-connection PRAGMA statements on each connection.
   ///  3. Creating custom SQLite functions.
   ///  4. Creating temporary views or triggers.
-  SqliteDatabaseImpl.withFactory(this.openFactory,
-      {this.maxReaders = SqliteDatabase.defaultMaxReaders}) {
-    updates = updatesController.stream;
-
-    _listenForEvents();
-
+  SqliteDatabaseImpl.withFactory(AbstractDefaultSqliteOpenFactory factory,
+      {this.maxReaders = SqliteDatabase.defaultMaxReaders})
+      : openFactory = factory as DefaultSqliteOpenFactory {
     _internalConnection = _openPrimaryConnection(debugName: 'sqlite-writer');
     _pool = SqliteConnectionPool(openFactory,
-        upstreamPort: _eventsPort.client(),
-        updates: updates,
         writeConnection: _internalConnection,
         debugName: 'sqlite',
         maxReaders: maxReaders,
         mutex: mutex);
-
-    isInitialized = _init();
-  }
-
-  Future<void> _init() async {
-    await _internalConnection.ready;
+    // Updates get updates from the pool
+    updates = _pool.updates;
   }
 
   @override
@@ -105,50 +93,6 @@ class SqliteDatabaseImpl
     return _pool.getAutoCommit();
   }
 
-  void _listenForEvents() {
-    UpdateNotification? updates;
-
-    Map<SendPort, StreamSubscription> subscriptions = {};
-
-    _eventsPort = PortServer((message) async {
-      if (message is UpdateNotification) {
-        if (updates == null) {
-          updates = message;
-          // Use the mutex to only send updates after the current transaction.
-          // Do take care to avoid getting a lock for each individual update -
-          // that could add massive performance overhead.
-          mutex.lock(() async {
-            if (updates != null) {
-              updatesController.add(updates!);
-              updates = null;
-            }
-          });
-        } else {
-          updates!.tables.addAll(message.tables);
-        }
-        return null;
-      } else if (message is InitDb) {
-        await isInitialized;
-        return null;
-      } else if (message is SubscribeToUpdates) {
-        if (subscriptions.containsKey(message.port)) {
-          return;
-        }
-        final subscription = updatesController.stream.listen((event) {
-          message.port.send(event);
-        });
-        subscriptions[message.port] = subscription;
-        return null;
-      } else if (message is UnsubscribeToUpdates) {
-        final subscription = subscriptions.remove(message.port);
-        subscription?.cancel();
-        return null;
-      } else {
-        throw ArgumentError('Unknown message type: $message');
-      }
-    });
-  }
-
   /// A connection factory that can be passed to different isolates.
   ///
   /// Use this to access the database in background isolates.
@@ -157,25 +101,13 @@ class SqliteDatabaseImpl
     return IsolateConnectionFactoryImpl(
         openFactory: openFactory,
         mutex: mutex.shared,
-        upstreamPort: _eventsPort.client());
-  }
-
-  SqliteConnectionImpl _openPrimaryConnection({String? debugName}) {
-    return SqliteConnectionImpl(
-        upstreamPort: _eventsPort.client(),
-        primary: true,
-        updates: updates,
-        debugName: debugName,
-        mutex: mutex,
-        readOnly: false,
-        openFactory: openFactory);
+        upstreamPort: _pool.upstreamPort!);
   }
 
   @override
   Future<void> close() async {
     await _pool.close();
     updatesController.close();
-    _eventsPort.close();
     await mutex.close();
   }
 
@@ -221,5 +153,14 @@ class SqliteDatabaseImpl
       {Duration? lockTimeout, String? debugContext}) {
     return _pool.writeLock(callback,
         lockTimeout: lockTimeout, debugContext: debugContext);
+  }
+
+  SqliteConnectionImpl _openPrimaryConnection({String? debugName}) {
+    return SqliteConnectionImpl(
+        primary: true,
+        debugName: debugName,
+        mutex: mutex,
+        readOnly: false,
+        openFactory: openFactory);
   }
 }
