+++ lib/src/sqlite_queries.dart
@@ -9,9 +9,6 @@ import 'update_notification.dart';
 /// Classes using this need to implement [SqliteConnection.readLock]
 /// and [SqliteConnection.writeLock].
 mixin SqliteQueries implements SqliteWriteContext, SqliteConnection {
-  /// Broadcast stream that is notified of any table updates
-  Stream<UpdateNotification>? get updates;
-
   @override
   Future<sqlite.ResultSet> execute(String sql,
       [List<Object?> parameters = const []]) async {
