+++ lib/src/web/worker/worker_utils.dart
@@ -1,5 +1,5 @@
 import 'package:sqlite_async/sqlite3_common.dart';
-import 'package:sqlite_async/src/web/database/web_db_context.dart';
+import 'package:sqlite_async/src/web/database/connection/drift_sqlite_connection.dart';
 
 void setupCommonWorkerDB(CommonDatabase database) {
   /// Exposes autocommit via a query function
